Update migrations and APIKeyGuard
