+++ packages/apps/job-launcher/server/src/common/guards/index.ts
@@ -1,3 +1,3 @@
 export * from './jwt.auth';
 export * from './signature.auth';
-export * from './jwt.apikey.auth';
+export * from './apikey.auth';
