+++ packages/apps/job-launcher/server/src/modules/auth/apikey.repository.ts
@@ -45,6 +45,5 @@ export class ApiKeyRepository {
       where: { id: apiKeyId },
       relations: ['user'],
     });
-  }
-  
+  }  
 }
