+++ packages/apps/job-launcher/server/src/modules/job/job.controller.ts
@@ -12,7 +12,7 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiQuery, ApiTags } from '@nestjs/swagger';
-import { JwtAuthGuard, JwtOrApiKeyGuard, SignatureAuthGuard } from '../../common/guards';
+import { JwtAuthGuard, ApiKeyGuard, SignatureAuthGuard } from '../../common/guards';
 import { RequestWithUser } from '../../common/types';
 import {
   JobFortuneDto,
@@ -37,7 +37,7 @@ import { Role } from '../../common/enums/role';
 export class JobController {
   constructor(private readonly jobService: JobService) {}
 
-  @UseGuards(JwtOrApiKeyGuard)
+  @UseGuards(ApiKeyGuard)
   @Post('/fortune')
   public async createFortuneJob(
     @Request() req: RequestWithUser,
