+++ packages/apps/job-launcher/server/src/modules/job/job.module.ts
@@ -11,12 +11,14 @@ import { JobRepository } from './job.repository';
 import { Web3Module } from '../web3/web3.module';
 import { RoutingProtocolService } from './routing-protocol.service';
 import { StorageModule } from '../storage/storage.module';
+import { AuthModule } from '../auth/auth.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([JobEntity]),
     ConfigModule,
     HttpModule,
+    AuthModule,
     PaymentModule,
     Web3Module,
     StorageModule
