+++ packages/apps/job-launcher/server/src/modules/user/user.entity.ts
@@ -41,7 +41,9 @@ export class UserEntity extends BaseEntity implements IUser {
   @OneToMany(() => PaymentEntity, (payment) => payment.user)
   public payments: PaymentEntity[];
 
-  @OneToOne(() => ApiKeyEntity, apiKey => apiKey.user)
+  @OneToOne(() => ApiKeyEntity, apiKey => apiKey.user, {
+    nullable: true,
+  })
   public apiKey: ApiKeyEntity;
 
 }
